requirements.txt
fastapi
uvicorn[standard]
groq
chromadb
pydantic
python-dotenv
tiktoken